+++ src/Profile/Magento/DataSelection/ProductReviewDataSelection.php
@@ -11,10 +11,12 @@ use Swag\MigrationMagento\Profile\Magento\DataSelection\DataSet\CrossSellingData
 use Swag\MigrationMagento\Profile\Magento\DataSelection\DataSet\CustomerDataSet;
 use Swag\MigrationMagento\Profile\Magento\DataSelection\DataSet\ManufacturerDataSet;
 use Swag\MigrationMagento\Profile\Magento\DataSelection\DataSet\ProductChildMultiSelectPropertyRelationDataSet;
+use Swag\MigrationMagento\Profile\Magento\DataSelection\DataSet\ProductChildMultiSelectTextPropertyRelationDataSet;
 use Swag\MigrationMagento\Profile\Magento\DataSelection\DataSet\ProductChildPropertyRelationDataSet;
 use Swag\MigrationMagento\Profile\Magento\DataSelection\DataSet\ProductCustomFieldDataSet;
 use Swag\MigrationMagento\Profile\Magento\DataSelection\DataSet\ProductDataSet;
 use Swag\MigrationMagento\Profile\Magento\DataSelection\DataSet\ProductMultiSelectPropertyRelationDataSet;
+use Swag\MigrationMagento\Profile\Magento\DataSelection\DataSet\ProductMultiSelectTextPropertyRelationDataSet;
 use Swag\MigrationMagento\Profile\Magento\DataSelection\DataSet\ProductOptionRelationDataSet;
 use Swag\MigrationMagento\Profile\Magento\DataSelection\DataSet\ProductPropertyRelationDataSet;
 use Swag\MigrationMagento\Profile\Magento\DataSelection\DataSet\ProductReviewDataSet;
@@ -56,7 +58,9 @@ class ProductReviewDataSelection implements DataSelectionInterface
             new ProductPropertyRelationDataSet(),
             new ProductChildPropertyRelationDataSet(),
             new ProductMultiSelectPropertyRelationDataSet(),
+            new ProductMultiSelectTextPropertyRelationDataSet(),
             new ProductChildMultiSelectPropertyRelationDataSet(),
+            new ProductChildMultiSelectTextPropertyRelationDataSet(),
             new ProductOptionRelationDataSet(),
             new CrossSellingDataSet(),
             new ProductReviewDataSet(),