+++ src/Profile/Magento19/Converter/Magento19ProductChildMultiSelectPropertyRelationConverter.php
@@ -9,6 +9,7 @@ namespace Swag\MigrationMagento\Profile\Magento19\Converter;
 
 use Swag\MigrationMagento\Profile\Magento\Converter\ProductMultiSelectPropertyRelationConverter;
 use Swag\MigrationMagento\Profile\Magento\DataSelection\DataSet\ProductChildMultiSelectPropertyRelationDataSet;
+use Swag\MigrationMagento\Profile\Magento\DataSelection\DataSet\ProductChildMultiSelectTextPropertyRelationDataSet;
 use Swag\MigrationMagento\Profile\Magento19\Magento19Profile;
 use SwagMigrationAssistant\Migration\MigrationContextInterface;
 
@@ -17,6 +18,9 @@ class Magento19ProductChildMultiSelectPropertyRelationConverter extends ProductM
     public function supports(MigrationContextInterface $migrationContext): bool
     {
         return $migrationContext->getProfile() instanceof Magento19Profile
-            && $this->getDataSetEntity($migrationContext) === ProductChildMultiSelectPropertyRelationDataSet::getEntity();
+            && (
+                $this->getDataSetEntity($migrationContext) === ProductChildMultiSelectPropertyRelationDataSet::getEntity()
+                || $this->getDataSetEntity($migrationContext) === ProductChildMultiSelectTextPropertyRelationDataSet::getEntity()
+            );
     }
 }